Add a MappingResolver to separate concerns
